+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Views/BottomToolbar/StreamingState/dotnet/CapturedImagesGalleryButton.cs
@@ -1,0 +1,8 @@
+namespace DIPS.Mobile.UI.API.Camera.ImageCapturing.Views.BottomToolbar.StreamingState;
+
+internal sealed partial class CapturedImagesGalleryButton
+{
+    private partial void SetThumbnailBlurred(bool shouldBlurThumbnail)
+    {
+    }
+}